CHECKING ALPHABETS  1
Checking alphabets
Lucy Pevensie and her brothers Peter and Edmund found themselves in Narnia,the land of magic during the World War III. Narnia was completely filled with gentle people and also it is where the trees sing, fauns dance and animals talk. Being from England, Lucy wanted to teach the kids , English language. She started teaching them alphabets but then she remembered that she might go back to London and felt sad. Edmund and Peter discussed with each other and suggested an idea to Lucy to come up with a program so that the kids can learn on their own when she was not there. Can you help Lucy to write a C++program to check whether the given character is a vowel or Consonant.
INPUT & OUTPUT FORMAT:

The input consists of a character.

The output should be any one of the below-given strings.

“Vowel” or “Consonant” or “Not an alphabet”.

SAMPLE INPUT & OUTPUT:

e

Vowel

z
Consonant
#coding
#include<iostream>
using namespace std;
int main()
{
  char n;
  cin >>n;
  if((n>='a'&&n<='z')||(n>='A'&&n<='Z'))
  {
    switch(n)
    {
      case 'a':
      case 'e':
      case 'i':
      case 'o':
      case 'u':
      case 'A':
      case 'E':
      case 'I':
      case 'O':
      case 'U':
        cout<<"Vowel";
        break;
      default:
        cout<<"Consonant";
        break;
    }
  }
  else
    cout<<"Not an alphabet";
  return 0;
}