DECISION MAKING - 1:BOATING
Boating
Tyrion Lannister ,Chandler Bing and Sheldon Cooper went on a tour to Zanzibar with their friends. As a part of the tour, you went boating with them. For the boat to remain stable, the number of people on one boat is restricted based on the weight of the people. You find that the boatman who is sailing your boat is so much greedy of money. For earning more, he takes too many people to travel in the boat at a time. So you want to check how many people can travel in the boat at a time so that the boat will not drown. Calculate the weight by considering the number of adults and number of children. Assume that an adult weighs 75 kg and children weigh 30 kg each. If the weight is normal, display Boat is stable, else display Boat will sink.
INPUT & OUTPUT FORMAT:

Input consists of 3 integers.

First input corresponds to the weight that the boat can handle.

Second input corresponds to the number of adults.
Third input corresponds to the number of children.

SAMPLE INPUT: 

340

2

3

SAMPLE OUTPUT: 
Boat is stable
#CODE
#include<iostream>
using namespace std;
int main()
{
  int a,b,c,adult,children,h;
  cin>>a>>b>>c;
  adult=b*75;
  children=c*30;
  h=adult+children;
  if(h<=a)
  {
     cout<<"Boat is stable";
  }
  else
  { 
    cout<<"Boat will sink";
  }
}