DECISION MAKING - 1:ELECTRICITY BILL
In the city King's Landing,the Golden company is taking care of the expenses for the Court's discussion room. Due to continuous discussion regarding the upcoming war with dragons the electricity bill was high where they need to pay it with all the money they had. Electricity board has decided to charge rupees based on the units consumed by a particular home. If the units consumed is less than or equal to 200, the cost for one unit is 0.5. If the unit is less than or equal to 400, the cost for one unit is 0.65 and Rs.100 extra charge. If the unit is less than or equal to 600, the cost for one unit is 0.80 and Rs.200 extra charge. If the unit is greater than 600 the cost for one unit is 1.25 and Rs.425 extra charge. You need to now calculate the electricity bill based on the units consumed (given input).
INPUT & OUTPUT FORMAT:

Input consists of one integer.

This corresponds to the units consumed.
The output consists of one integer.

This corresponds to the electricity bill in rupees.

SAMPLE INPUT:

200

SAMPLE OUTPUT:

Rs.100
#CODE
#include<iostream>
using namespace std;
int main()
{
  int unit,amt;
  cin>>unit;
  if(unit<=200)
    amt = unit*0.5;
  if(unit>200 && unit<=400)
    amt = amt + (unit-200)*0.65 + 100;
  if(unit>400 && unit<=600)
    amt = amt +(unit-400)*0.80 + 200;
  if(unit>600)
    amt = amt + (unit-600)*1.25 + 425;
  cout<<"Rs."<<amt;
}
 O/P:
IN:210
PO:Rs:236