DECISION MAKING 1 - GARDENING
Gardening
Arya Stark of Winterfell is very much interested in gardening and she decides to plant more trees in her garden. She plants trees in the order of rows and columns. She numbered the trees in column wise order. She planted the orange trees only in the second column from both first and last. But later when the trees grew up, she forgot where she planted orange trees. So help her find out whether the given tree number is a number of orange trees or not. Display whether “It is an orange tree” or “It is not an orange tree”.

INPUT FORMAT:
nput consists of 3 integers.

First input corresponds to the number of rows.

Second input corresponds to the number of columns.

Third input corresponds to the tree number.

SAMPLE INPUT:

5

5

20

SAMPLE OUTPUT:

It is an orange tree


Case 1
Input (stdin)
10
10
21

Output (stdout)
It is not an orange tree
#CODE
#include<iostream>
using namespace std;
int main()
{
  int rows,coloums,tree;
  cin >> rows >> coloums >> tree;
  if(rows+coloums==tree|| (rows+coloums)*2==tree)
     cout <<  "It is an orange tree";
  else
    cout << "It is not an orange tree";
  
}