DECISION MAKING 1- LAB ALLOCATION 3
Lab Allocation - III
There are 3 labs in the Chemical department(L1, L2, and L3) with a seating capacity of x, y, and z. A single lab needs to be allocated to a class of 'n' students. How many of the 3 labs can accommodate 'n' students?

INPUT FORMAT:

Input consists of 4 integers.

The first input is an integer which denotes the seating capacity of L1(x).

The second input is an integer which denotes the seating capacity of L2(y).

The third input is an integer which denotes the seating capacity of L3(z).

The fourth input is an integer which denotes the number of students(n).

OUTPUT FORMAT:

Print the number of labs which can accommodate the 'n' number of students.

Refer the sample output for formatting.

SAMPLE INPUT:

30

40

20

25

SAMPLE OUTPUT:

2 
#CODE
#include<iostream>
using namespace std;
int main()
{
  int l1,l2,l3,n,count=0;
  cin >> l1 >> l2 >> l3 >> n;
  if(l1>=n)
    count++;
  if(l2>=n)
    count++;
  if(l3>=n)
    count++;
  cout<<count;
  return 0;
}