DECISION MAKING 1 - PAINTING
Paintings
A wall is of length L and breadth B. Leonardo Da Vinci is trying to fit two rectangular paintings in it. He wants to fit these paintings on the wall such that they do not overlap on each other or they should not occupy space outside the area of the wall. He also knows the length and breadth of both the paintings. Help him in finding whether the two paintings can fit onto the wall properly or not.

INPUT & OUTPUT FORMAT:

Input consists of 6 integers.

First and second input corresponds to the length and breadth of the wall.

Third and fourth input corresponds to the length and breadth of the first painting.

Fifth and sixth input corresponds to the length and breadth of the second painting.

If both the paintings can be fixed on the wall, display “Leonardo can fix both painting” or display “Leonardo cannot fix both painting”.

SAMPLE INPUT:

8

8

3

2

4

4

SAMPLE OUTPUT:

Leonardo can fix both painting


Case 1
Case 2
Case 3
Input (stdin)
8
8
3
2
4
4

Output (stdout)
Leonardo can fix both painting
#CODE
#include<iostream>
using namespace std;
int main()
{
  int wl,wb,l1,b1,l2,b2,x,y,z,a;
  cin >> wl >> wb >> l1 >> l2 >> b1 >> b2;
  x=wl*wb;
  y=l1*b1;
  z=l2*b2;
  a=y+z;
  if(x>=a)
    cout << "Leonardo can fix both painting";
  else
    cout << "Leonardo cannot fix both painting";
  return 0;
}