DECISION MAKING 2 - ONLINE SHOPPING
Online shopping
Richard Castle wants to buy a shirt. As he is very lazy, he decided to buy the shirt online. He chooses a shirt in Flipkart and is surprised to see the same shirt in Amazon and Snapdeal as well. So he decided to buy the shirt from the website which offers it at the least price. The price of the shirt, discount % and the shipping charges of all three websites have been given as input. Help him in calculating the price of the shirt in each website and decide which website has the lowest price.

INPUT & OUTPUT FORMAT:

Input consist of 9 integers.

First three input corresponds to Flipkart details such as the amount of the shirt, discount offered and shipping charges.

Next three input corresponds to Snapdeal details such as the amount of shirt, discount offered and shipping charge.

Last three input corresponds to amazon details such as the amount of shirt, discount offered and shipping charge.

SAMPLE INPUT:

1000

50

50

900

50

70

800

10

200

SAMPLE OUTPUT:

In Flipkart: Rs.550

In Snapdeal: Rs.520

In Amazon: Rs.920

He will prefer Snapdeal

#CODE
#include<iostream>
using namespace std;
int main()
{
  int fa,fd,fs,sa,sd,ss,aa,ad,as,tf,ta,ts;
  cin>>fa>>fd>>fs>>sa>>sd>>ss>>aa>>ad>>as;
  tf=fa-fd*fa/100.0+fs;
  ta=aa-ad*aa/100.0+as;
  ts=sa-sd*sa/100.0+ss;
  cout <<"In Flipkart Rs."<<tf<<endl;
  cout <<"In Snapdeal Rs."<<ts<<endl;
  cout <<"In Amazon Rs."<<ta<<endl;
  if(tf<=ta && tf<=ts)
    cout <<"He will prefer Flipkart";
  else if(ta<=tf && ta<=ts)
    cout <<"He will prefer Amazon";
  else
    cout <<"He will prefer Snapdeal";
  return 0;
  }