DECISION MAKING 1- GAME WITH SHAPES
Game with shapes
Sherlock Holmes loves to play with shapes and his sister Eurus Holmes is interested in Mathematics. They got a wooden box which is square in shape with length 'x' cm and a round bangle with radius 'r' cm and decided to play with it. They started playing and his sister asked him to find whether the circle will fit into the square. Help Sherlock Holmes to find whether the circle will fit into a square or not. If the circle fits, display “Circle can be inside a square” else display “Circle cannot be inside a square”.

INPUT FORMAT:

Input consist of 2 integers.

First input corresponds to the radius of a circle.

Second input corresponds to the length of a square.

SAMPLE INPUT: 
77

65

SAMPLE OUTPUT: 

Circle cannot be inside a square
#CODE
#include<iostream>
using namespace std;
int main()
{
  int r,l,x,A,d;
  cin>>r>>l;
  A=3.14*r*r;
  x=l*l;
  d=2*r;
  if(A<x && d<=l)
    cout<<"Circle can be inside a square";
  else 
    cout<<"Circle cannot be inside a square";
  return 0;
 
}
Input (stdin)
65
45

Output (stdout)
Circle cannot be inside a square