DECISION MAKING 1 - LAB ALLOCATION 2
Lab Allocation - II
There are 3 labs in the Chemical department(L1, L2, and L3) with a seating capacity of x, y, and z respectively. One of the 3 labs has been allocated for FACE training. Out of the available labs(after allocating one to FACE training), write a program to find the lab which has minimal seating capacity.

INPUT FORMAT:

Input consists of 3 integers and a string.

The first input is an integer which denotes the seating capacity of L1(x).

The second input is an integer which denotes the seating capacity of L2(y).

The third input is an integer which denotes the seating capacity of L3(z).

The fourth input is a string which denotes the lab which is allocated for FACE training.

OUTPUT FORMAT:

Print the minimal seating capacity lab among the available labs.

Refer the sample output for formatting.

SAMPLE INPUT:

30

40

20

L3

SAMPLE OUTPUT:

L1
#CODE
#include<iostream>
using namespace std;
int main()
{
  int x,y,z;
  string str;
  cin >>x>>y>>z;
  cin>>str;
  if (str=="L1")
  { if(y<z)
    cout<<"L2";
    else
    cout<<"L3";
   }
   else if(str=="L2")
  {
     if(x<z)
     cout<<"L1";
     else
     cout<"L3";
  }
  else 
  {  
    if(y<x)
    cout<<"L2";
    else
     cout<<"L1";
  }
  return 0;
}