1D ARRAY - REMOVE DUPLICATE ELEMENTS
Remove duplicate elements
Sherlock and John were investigating a murder and the murderer was Ms.Irene Adler. She wanted Sherlock to find her by solving the puzzles given by her and the tasks given by her. Sherlock was able to solve all but one. The last task was to write a C++ program to remove the duplicate elements from an array. Can you help Sherlock?

INPUT FORMAT:

Input consists of 1 integer and 1 array.

The first integer corresponds to the size of the array.

The next integers correspond to the elements in the array.

OUTPUT FORMAT:

The output consists of an array with no duplicate elements.

SAMPLE INPUT:

3

 1

 5

 1

SAMPLE OUTPUT:

 1

 5
#CODE
#include<iostream>
using namespace std;
int main()
{
  int n,i,j;
  cin>>n;
  int arr[n];
  for(i=0;i<n;i++)
    cin>>arr[i];
  
  cout<<arr[0]<<endl;
  for(i=1;i<n;i++)
  {
    for(j=0;j<i;j++)
    {
      if (arr[j]==arr[i])
        break;
    }
    if(j==i)
      cout<<arr[i]<<endl;
  }
return 0;
}