RECURSION - SUM OF EVEN NUMBERS
Sum of even numbers
Dolores Abarnathy and Maeve Millay were playing a puzzle game with a given set of numbers. They need to find the even numbers and find the sum of the even numbers. Write a program to help them to solve the puzzle game and to win the mobile phone.

Input Format:

Input consists of n+1 integers. 

he first integer corresponds to ‘n’, the size of the array. 

The next ‘n’ integers correspond to the elements in the array. 

Assume that the maximum value of n is 15.

Output Format:

Refer to sample output for details.

Sample Input:

5

2

3

6

8

-1

Sample Output:

Enter the number of elements in the array

Enter the elements in the array

The sum of the even elements in the array is 16

#CODE
#include<iostream>
using namespace std;
int sumeven(int arr[],int pos,int n,int sum)
{
  if(pos==n)
    return sum;
  else if(arr[pos]%2==0)
    return sumeven(arr,pos+1,n,sum+arr[pos]);
    else
      return sumeven(arr,pos+1,n,sum);
 }
int main()
{
  int n;
  cout<< "Enter the number of elements in the array" << endl;
  cin >> n;
  int arr[n];
  cout<< "Enter the elements in the array" << endl; 
  for(int i=0;i<=n;i++)
    cin >> arr[i];
  cout<< "The sum of the even elements in the array is " << sumeven(arr,0,n,0);
  return 0;
}